+++ Core/Inc/usart.h
@@ -32,9 +32,15 @@ extern "C" {
 
 /* USER CODE END Includes */
 
+extern UART_HandleTypeDef huart4;
+
 extern UART_HandleTypeDef huart5;
 
 extern UART_HandleTypeDef huart7;
+
+extern UART_HandleTypeDef huart8;
+
+extern UART_HandleTypeDef huart9;
 
 extern UART_HandleTypeDef huart1;
 
@@ -42,17 +48,23 @@ extern UART_HandleTypeDef huart2;
 
 extern UART_HandleTypeDef huart3;
 
+extern UART_HandleTypeDef huart6;
+
 extern UART_HandleTypeDef huart10;
 
 /* USER CODE BEGIN Private defines */
 
 /* USER CODE END Private defines */
 
+void MX_UART4_Init(void);
 void MX_UART5_Init(void);
 void MX_UART7_Init(void);
+void MX_UART8_Init(void);
+void MX_UART9_Init(void);
 void MX_USART1_UART_Init(void);
 void MX_USART2_UART_Init(void);
 void MX_USART3_UART_Init(void);
+void MX_USART6_UART_Init(void);
 void MX_USART10_UART_Init(void);
 
 /* USER CODE BEGIN Prototypes */