+++ Core/Src/main.c
@@ -138,6 +138,10 @@ int main(void)
   MX_TIM8_Init();
   MX_TIM12_Init();
   MX_ADC1_Init();
+  MX_USART6_UART_Init();
+  MX_UART4_Init();
+  MX_UART8_Init();
+  MX_UART9_Init();
   /* USER CODE BEGIN 2 */
   System_Init();
   /* USER CODE END 2 */
@@ -280,27 +284,27 @@ void MPU_Config(void)
 
 }
 
-/**
-  * @brief  Period elapsed callback in non blocking mode
-  * @note   This function is called  when TIM2 interrupt took place, inside
-  * HAL_TIM_IRQHandler(). It makes a direct call to HAL_IncTick() to increment
-  * a global variable "uwTick" used as application time base.
-  * @param  htim : TIM handle
-  * @retval None
-  */
-void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim)
-{
-  /* USER CODE BEGIN Callback 0 */
+// /**
+//   * @brief  Period elapsed callback in non blocking mode
+//   * @note   This function is called  when TIM2 interrupt took place, inside
+//   * HAL_TIM_IRQHandler(). It makes a direct call to HAL_IncTick() to increment
+//   * a global variable "uwTick" used as application time base.
+//   * @param  htim : TIM handle
+//   * @retval None
+//   */
+// void HAL_TIM_PeriodElapsedCallback(TIM_HandleTypeDef *htim)
+// {
+//   /* USER CODE BEGIN Callback 0 */
 
-  /* USER CODE END Callback 0 */
-  if (htim->Instance == TIM2)
-  {
-    HAL_IncTick();
-  }
-  /* USER CODE BEGIN Callback 1 */
+//   /* USER CODE END Callback 0 */
+//   if (htim->Instance == TIM2)
+//   {
+//     HAL_IncTick();
+//   }
+//   /* USER CODE BEGIN Callback 1 */
 
-  /* USER CODE END Callback 1 */
-}
+//   /* USER CODE END Callback 1 */
+// }
 
 /**
   * @brief  This function is executed in case of error occurrence.