+++ User_File/Task/CanTxTask.cpp
@@ -28,9 +28,8 @@ extern "C" {
 
 extern "C" void Can_Tx_Task(void *argument)
 {
-
+  TickType_t xLastWakeTime = xTaskGetTickCount();
   for (;;) {
-    TickType_t xLastWakeTime = xTaskGetTickCount();
     vTaskDelayUntil(&xLastWakeTime, pdMS_TO_TICKS(1)); // 每1ms发送一次CAN消息,根据需要调整频率
     BSP_CAN_SendAsync();
     BSP_CAN_SendPer();